src/main/java/com/mediscreen/patient/controller/ExceptionHandlerController.java
package com.mediscreen.patient.controller;

import com.mediscreen.patient.exception.PatientNotFoundException;
import org.apache.logging.log4j.LogManager;
import org.apache.logging.log4j.Logger;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.ControllerAdvice;
import org.springframework.web.bind.annotation.ExceptionHandler;

@ControllerAdvice
public class ExceptionHandlerController {


    private static final Logger log = LogManager.getLogger(ExceptionHandlerController.class);

    /**
     * Handle PatientNotFoundException response entity.
     * @param ex the ex
     * @return the response entity
     */
    @ExceptionHandler(PatientNotFoundException.class)
    public ResponseEntity<String> handlePatientNotFoundException(PatientNotFoundException ex) {
        log.error("PatientNotFoundException", ex);
        return ResponseEntity.status(HttpStatus.NOT_FOUND).body(ex.getMessage());
    }
}
